Add files via upload
Updated data scrapper, not finished currently working on debugging on the program scrapes through each link on the page and moves to the next page, scrapes the links on that site until there are no more links or pages to search (ideally). I also am working on trying to print out the keywords.